+++ CHANGES.md
@@ -9,6 +9,7 @@ adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Support for Raspberry Pi (Linux)
+- Partial support for windows
 - Jest test suite and tests for Linux and other platforms
   
 ### Changed